refactor(patchers): move EngineDetectorResult to Contracts, wire IEngineDetector
- EngineDetectorResult moved to TombLauncher.Contracts.EngineDetectors
- IEngineDetector interface defined in Contracts with Detect() method
- TombRaiderEngineDetector now implements IEngineDetector
- IPatcher.EngineDetector typed as IEngineDetector (no circular dependency)
- WidescreenPatcher updated accordingly
