+++ src/TombLauncher.Contracts/EngineDetectors/IEngineDetector.cs
@@ -1,0 +1,6 @@
+namespace TombLauncher.Contracts.EngineDetectors;
+
+public interface IEngineDetector
+{
+    EngineDetectorResult Detect(string containingFolder);
+}
