+++ src/TombLauncher.Contracts/Patchers/IPatcher.cs
@@ -1,8 +1,10 @@
+using TombLauncher.Contracts.EngineDetectors;
+
 namespace TombLauncher.Contracts.Patchers;
 
 public interface IPatcher
 {
     Task<PatchResult> DetectChanges(string targetFolder);
     Task<PatchResult> ApplyPatch(string targetFolder, IPatchParameters parameters);
-    TombRaiderEngineDetector EngineDetector { get; set; }
+    IEngineDetector EngineDetector { get; set; }
 }
